requirements.txt
speechrecognition
pvporcupine
pyaudio
requests
playsound==1.2.2
openai>=1.3.8
playsound
requests
speechrecognition
pvporcupine
tk  # (optional, comes with most Python installs)
PyQt6
pythreejs
chromadb
tiktoken
python-dotenv
pygame
torch 
torchvision 
torchaudio
git+https://github.com/openai/whisper.git
pyaudio
webrtcvad
resampy
fastapi
uvicorn[standard]
python-socketio[asyncio] 
aiofiles
pyttsx3
pywin32
